requirements.txt
aiosqlite==0.20.0
amqp==5.3.1
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
connexion==3.1.0
cryptography==43.0.3
ecdsa==0.19.0
Flask==3.1.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
MarkupSafe==3.0.2
packaging==24.2
prompt_toolkit==3.0.48
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
uvicorn-worker==0.2.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.1.3
